--- a/app/src/utils/utils.js
+++ b/app/src/utils/utils.js
@@ -116,32 +116,19 @@ const getStartOfTheDay = () => {
   return today.toISOString()
 }
 
-// Core i5 2.9GHz
-// 4 arrays with 90 elements
-// No-throttle: mergeTimeseriesByKey-old: ~4.4ms
-// No-throttle: mergeTimeseriesByKey-new: ~0.3ms
-
-// x4 throttle: mergeTimeseriesByKey-old: ~13.2ms
-// x4 throttle: mergeTimeseriesByKey-new: ~1.3ms
-
-// x6 throttle: mergeTimeseriesByKey-old: ~19.7
-// x6 throttle: mergeTimeseriesByKey-new: ~2.1ms
-
 const mergeTimeseriesByKey = ({ timeseries, key: mergeKey }) => {
-  const timeseriesLength = timeseries.length
-
-  let longestTSIndex = 0
-  for (let i = 1; i < timeseriesLength; i++) {
-    if (timeseries[longestTSIndex].length < timeseries[i].length) {
-      longestTSIndex = i
-    }
-  }
-
-  const [mutableLongestTS] = timeseries.splice(longestTSIndex, 1)
-  const longestTS = mutableLongestTS.slice()
+  const longestTS = timeseries
+    .reduce((acc, val) => {
+      return acc.length > val.length ? acc : val
+    }, [])
+    .slice()
   const longestTSLastIndex = longestTS.length - 1
 
   for (const timeserie of timeseries) {
+    if (timeserie === longestTS) {
+      continue
+    }
+
     let longestTSRightIndexBoundary = longestTSLastIndex
 
     for (