#### Summary
This PR increases calculation speed of the `mergeTimeseriesByKey` algorithm.

### Setup
- **Specs**: Core i5 2.9GHz
- **Data**: 4 arrays with 90 elements
- Throttling was achieved by using Google Chrome Dev Tools

### Benchmark results
| | Old algorithm | New algorithm |
| :---         |     :---:      |          :---: |
| No throttling   | ~4.4ms     | ~0.3ms    |
| 4x slowdown     | ~13.2ms       | ~1.3ms      |
| 6x slowdown     | ~19.7ms      | ~2.1ms      |